Automated g4 rollback of changelist 130106966.
*** Reason for rollback ***

breaks stuff that should be using the binding SPI

*** Original change description ***

Move the implementation classes from the multibindings package to inject/internal.

Now that the multibindings package is part of the main guice project we can do things like
* install the MultibindingsScanner by default
* rewrite multibinders to use InternalFactory objcts

But for those to work they need to be in the internal package.

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=130114909