Add getEntries(Iterable) to MapBinderBinding
This change is motivated by ModuleTester.assertMappedKey.
Previously there was no good way to determine the K keys
that were actually bound in the MapBinder. ModuleTester
relied upon implementation details in RealMapBinder to
do the assert, but that is not good. By providing this
new method, it will provide ModuleTester, and others, a
better way to investigate MapBindings.

A change to ModuleTester to use this new method will
follow in a later CL.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=134287856