Add API to allow custom error message formatting.
Since `Message` is used in various public APIs it's infesiable to change
it to be a subclass of a error message class. Instead this CL moves all the
details about the error into an `ErrorDetail` class and wraps it with `Message`. Errors that needs custom formatting can inherit `ErrorDetail` and a generic implementation is added that uses the current formatting options.

Follow up changes to add different implementations for specific errors that should be formatted differently if the experimental new error feature is enabled.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=318520440