Add a check for OptionalBinder infinite recursion.
When an optional binder binding was linked to the same class like below, infinite recursion occurs. There is a check for recursive binding in BindingProcessor class for all linked bindings but that did not work for OptionalBinder since OptionalBinder original key annotation is wrapped into something. So, comparing the original key and linked binding key didn't work.

Added a check for that.

Example:

```
OptionalBinder.newOptionalBinder(binder(), MyClass.class)
    .setBinding()
    .to(MyClass.class);
```

PiperOrigin-RevId: 332062297